requirements/requirements.txt
# Основные зависимости
qdrant-client>=1.6.4
sentence-transformers>=2.2.2
spacy>=3.7.2
pymupdf>=1.23.17
python-docx>=1.1.0
tqdm>=4.66.1
PyYAML>=6.0.1
pytesseract>=0.3.10
Pillow>=10.0.1
httpx>=0.24.1
fastapi>=0.95.2
uvicorn>=0.22.0
Jinja2>=3.1.2
pandas>=1.5.3
numpy>=1.24.3
scikit-learn>=1.3.0
scipy>=1.10.1  # Добавлено для MMR
nltk>=3.8.1     # Добавлено для очистки текста

# Удалены:
# camelot-py[base]>=0.11.0 - больше не используется
# concurrent-log-handler>=0.9.24 - заменено стандартным logging
# transformers>=4.34.0 - не используется напрямую
# sentencepiece>=0.1.99 - не требуется